Updated unsigned int to unsigned long long in c tests

+++ tests/test_fa_get_all_results.c
@@ -19,7 +19,7 @@ int main(int argc, const char* argv[])
     // read parameters
     tone_type ttype;
     int qres;
-    unsigned int npts, navg, nfft, tmp_win, num_tones;
+    unsigned long long npts, navg, nfft, tmp_win, num_tones;
     double *freq;
     GnWindow win;    
     err_code = read_scalar_from_json_file(test_filename, "wf_type", (void*)(&ttype), UINT64);
