+++ tests/test_fft.c
@@ -16,7 +16,7 @@ int main(int argc, const char* argv[])
     // read parameters
     tone_type ttype;
     int qres;
-    unsigned int npts, fft_navg, nfft, tmp_win;
+    unsigned long long npts, fft_navg, nfft, tmp_win;
     GnWindow win;
     err_code = read_scalar_from_json_file(test_filename, "wf_type", (void*)(&ttype), UINT64);
     err_code = read_scalar_from_json_file(test_filename, "qres", (void*)(&qres), INT32);
